Now using 32-bits size description, since Ted sends some huge messages that got truncated when using 16-bits description.

+++ include/CoClient.h
@@ -105,7 +105,7 @@ private:
 	miString clientType;
 	miString lockFile, serverCommand, host;
 
-	quint16 blockSize;
+	quint32 blockSize;
 	quint16 port;
 	QProcess *server;
